+++ lib/eth/client.rb
@@ -1,4 +1,4 @@
-# Copyright (c) 2016-2022 The Ruby-Eth Contributors
+# Copyright (c) 2016-2023 The Ruby-Eth Contributors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -106,8 +106,8 @@ module Eth
     # See {#transfer} for params and overloads.
     #
     # @return [String] the transaction hash once it is mined.
-    def transfer_and_wait(destination, amount, sender_key = nil, legacy = false)
-      wait_for_tx(transfer(destination, amount, sender_key, legacy))
+    def transfer_and_wait(destination, amount, **kwargs)
+      wait_for_tx(transfer(destination, amount, **kwargs))
     end
 
     # Simply transfer Ether to an account without any call data or
@@ -117,19 +117,24 @@ module Eth
     # **Note**, that many remote providers (e.g., Infura) do not provide
     # any accounts. Provide a `sender_key` if you experience issues.
     #
-    # @param destination [Eth::Address] the destination address.
-    # @param amount [Integer] the transfer amount in Wei.
-    # @param sender_key [Eth::Key] the sender private key.
-    # @param legacy [Boolean] enables legacy transactions (pre-EIP-1559).
+    # @overload transfer(destination, amount)
+    #   @param destination [Eth::Address] the destination address.
+    #   @param amount [Integer] the transfer amount in Wei.
+    # @overload transfer(destination, amount, **kwargs)
+    #   @param destination [Eth::Address] the destination address.
+    #   @param amount [Integer] the transfer amount in Wei.
+    #   @param **sender_key [Eth::Key] the sender private key.
+    #   @param **legacy [Boolean] enables legacy transactions (pre-EIP-1559).
+    #   @param **nonce [Integer] optional specific nonce for transaction.
     # @return [String] the local transaction hash.
-    def transfer(destination, amount, sender_key = nil, legacy = false)
+    def transfer(destination, amount, **kwargs)
       params = {
         value: amount,
         to: destination,
         gas_limit: gas_limit,
         chain_id: chain_id,
       }
-      if legacy
+      if kwargs[:legacy]
         params.merge!({
           gas_price: max_fee_per_gas,
         })
@@ -139,22 +144,22 @@ module Eth
           max_gas_fee: max_fee_per_gas,
         })
       end
-      unless sender_key.nil?
+      unless kwargs[:sender_key].nil?
 
         # use the provided key as sender and signer
         params.merge!({
-          from: sender_key.address,
-          nonce: get_nonce(sender_key.address),
+          from: kwargs[:sender_key].address,
+          nonce: kwargs[:nonce] || get_nonce(kwargs[:sender_key].address),
         })
         tx = Eth::Tx.new(params)
-        tx.sign sender_key
+        tx.sign kwargs[:sender_key]
         return eth_send_raw_transaction(tx.hex)["result"]
       else
 
         # use the default account as sender and external signer
         params.merge!({
           from: default_account,
-          nonce: get_nonce(default_account),
+          nonce: kwargs[:nonce] || get_nonce(default_account),
         })
         return eth_send_transaction(params)["result"]
       end
@@ -189,6 +194,7 @@ module Eth
     #   @param **sender_key [Eth::Key] the sender private key.
     #   @param **legacy [Boolean] enables legacy transactions (pre-EIP-1559).
     #   @param **gas_limit [Integer] optional gas limit override for deploying the contract.
+    #   @param **nonce [Integer] optional specific nonce for transaction.
     # @return [String] the transaction hash.
     # @raise [ArgumentError] in case the contract does not have any source.
     def deploy(contract, *args, **kwargs)
@@ -223,7 +229,7 @@ module Eth
         # Uses the provided key as sender and signer
         params.merge!({
           from: kwargs[:sender_key].address,
-          nonce: get_nonce(kwargs[:sender_key].address),
+          nonce: kwargs[:nonce] || get_nonce(kwargs[:sender_key].address),
         })
         tx = Eth::Tx.new(params)
         tx.sign kwargs[:sender_key]
@@ -232,7 +238,7 @@ module Eth
         # Uses the default account as sender and external signer
         params.merge!({
           from: default_account,
-          nonce: get_nonce(default_account),
+          nonce: kwargs[:nonce] || get_nonce(default_account),
         })
         return eth_send_transaction(params)["result"]
       end
@@ -288,6 +294,8 @@ module Eth
     #   @param **legacy [Boolean] enables legacy transactions (pre-EIP-1559).
     #   @param **address [Eth::Address] contract address.
     #   @param **gas_limit [Integer] optional gas limit override for deploying the contract.
+    #   @param **nonce [Integer] optional specific nonce for transaction.
+    #   @param **tx_value [Integer] optional transaction value field filling.
     # @return [Object] returns the result of the transaction.
     def transact(contract, function, *args, **kwargs)
       gas_limit = if kwargs[:gas_limit]
@@ -297,7 +305,7 @@ module Eth
         end
       fun = contract.functions.select { |func| func.name == function }[0]
       params = {
-        value: 0,
+        value: kwargs[:tx_value] || 0,
         gas_limit: gas_limit,
         chain_id: chain_id,
         to: kwargs[:address] || contract.address,
@@ -317,7 +325,7 @@ module Eth
         # use the provided key as sender and signer
         params.merge!({
           from: kwargs[:sender_key].address,
-          nonce: get_nonce(kwargs[:sender_key].address),
+          nonce: kwargs[:nonce] || get_nonce(kwargs[:sender_key].address),
         })
         tx = Eth::Tx.new(params)
         tx.sign kwargs[:sender_key]
@@ -326,7 +334,7 @@ module Eth
         # use the default account as sender and external signer
         params.merge!({
           from: default_account,
-          nonce: get_nonce(default_account),
+          nonce: kwargs[:nonce] || get_nonce(default_account),
         })
         return eth_send_transaction(params)["result"]
       end
@@ -354,9 +362,9 @@ module Eth
     # @raise [ArgumentError] in case the contract cannot be called yet.
     def is_valid_signature(contract, hash, signature, magic = "1626ba7e")
       raise ArgumentError, "Contract not deployed yet." if contract.address.nil?
-      hash = Util.hex_to_bin hash if Util.is_hex? hash
-      signature = Util.hex_to_bin signature if Util.is_hex? signature
-      magic = Util.hex_to_bin magic if Util.is_hex? magic
+      hash = Util.hex_to_bin hash if Util.hex? hash
+      signature = Util.hex_to_bin signature if Util.hex? signature
+      magic = Util.hex_to_bin magic if Util.hex? magic
       result = call(contract, "isValidSignature", hash, signature)
       return result === magic
     end
@@ -373,7 +381,7 @@ module Eth
     #
     # @param hash [String] the transaction hash.
     # @return [Boolean] true if included in a block.
-    def tx_mined?(hash)
+    def mined?(hash)
       mined_tx = eth_get_transaction_by_hash hash
       !mined_tx.nil? && !mined_tx["result"].nil? && !mined_tx["result"]["blockNumber"].nil?
     end
@@ -398,7 +406,7 @@ module Eth
       retry_rate = 0.1
       loop do
         raise Timeout::Error if ((Time.now - start_time) > timeout)
-        return hash if tx_mined? hash
+        return hash if mined? hash
         sleep retry_rate
       end
     end
@@ -471,7 +479,7 @@ module Eth
         return Util.prefix_hex "#{params.to_i.to_s(16)}"
       elsif params.is_a? Address
         return params.to_s
-      elsif Util.is_hex? params
+      elsif Util.hex? params
         return Util.prefix_hex params
       else
         return params